refactor(recovermw): rename to recoverymw and clarify partial-write behavior
Address PR #1640 review feedback:
- Rename `internal/infrastructure/recover` to `recoverymw` to avoid shadowing
  Go's builtin recover() (per Claude review). Drops the per-call import alias
  in workers since the package name now matches.
- Rewrite the misleading "just bail out" comment on the deferred header
  write (per Gemini + Claude review): the middleware deliberately attempts
  to write a JSON error even when headers are already committed, knowing
  WriteHeader will be a no-op in that case. Document this as an explicit
  best-effort tradeoff over Cloudflare's CORS-less 1101 page.
- Add TestMiddleware_RecoversPanicAfterPartialWrite covering the
  panic-after-partial-write path so the documented limitation is enforced
  by a regression test.

No behavior change in the panic-recovery path itself.
