+++ cmd/workers/solo/main.go
@@ -12,7 +12,7 @@ import (
 	corsmw "github.com/yuta-yoshinaga/go_trumpcards/internal/infrastructure/cors"
 	"github.com/yuta-yoshinaga/go_trumpcards/internal/infrastructure/games"
 	_ "github.com/yuta-yoshinaga/go_trumpcards/internal/infrastructure/games/solo"
-	recovermw "github.com/yuta-yoshinaga/go_trumpcards/internal/infrastructure/recover"
+	"github.com/yuta-yoshinaga/go_trumpcards/internal/infrastructure/recoverymw"
 )
 
 func main() {
@@ -25,7 +25,7 @@ func main() {
 	// Wrap the mux with panic recovery so any handler panic produces a JSON
 	// 500 instead of Cloudflare's generic 1101 HTML error page (which loses
 	// CORS headers and shows up as "通信エラー" with no actionable detail).
-	var handler http.Handler = recovermw.Middleware(mux)
+	var handler http.Handler = recoverymw.Middleware(mux)
 	if origins := corsmw.ParseOrigins(cloudflare.Getenv("CORS_ALLOWED_ORIGINS")); origins != nil {
 		handler = corsmw.Middleware(origins, handler)
 	}
